카테고리 추가 컨트롤러 및 서비스 추가

+++ src/main/java/ac/su/kiosk/domain/Menu.java
@@ -16,7 +16,7 @@ public class Menu {
     private String name;
 
     @Column
-    private Double basePrice;
+    private Long basePrice;
 
     @Column
     private String description;
